rag-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
openai==1.78.1
python-multipart==0.0.20
python-dotenv==1.1.0
boto3==1.38.22
